Restructure
ExtractFromDirectory to demo, ResourceUtils to utils, new EidosAttachment in attachments and changed packages and imports to match
