+++ src/main/scala/org/clulab/wm/eidos/EidosActions.scala
@@ -3,6 +3,7 @@ package org.clulab.wm.eidos
 import java.io.File
 
 import com.typesafe.scalalogging.LazyLogging
+import org.clulab.wm.eidos.attachments._
 import org.clulab.odin._
 import org.clulab.odin.impl.Taxonomy
 import org.clulab.wm.eidos.Aliases.Quantifier
@@ -16,12 +17,6 @@ import scala.io.BufferedSource
 // 2) the methods available on the `State`
 
 //TODO: need to add polarity flipping
-
-
-case class Quantification(quantifier: Quantifier, adverbs: Option[Seq[String]]) extends Attachment
-case class Increase(trigger: String, quantifier: Option[Seq[Quantifier]] = None) extends Attachment
-case class Decrease(trigger: String, quantifier: Option[Seq[Quantifier]] = None) extends Attachment
-
 
 class EidosActions extends Actions with LazyLogging {
