+++ src/main/scala/org/clulab/wm/eidos/demo/ExtractFromDirectory.scala
@@ -1,12 +1,11 @@
-package org.clulab.wm.eidos
+package org.clulab.wm.eidos.demo
 
 import java.io.PrintWriter
-
 import scala.collection.Seq
-
 import org.clulab.serialization.json.stringify
 import org.clulab.wm.eidos.serialization.json.JLDCorpus
 import org.clulab.wm.eidos.utils.FileUtils.findFiles
+import org.clulab.wm.eidos.EidosSystem
 
 
 object ExtractFromDirectory extends App {
